pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "handwritten-grader"
version = "0.1.0"
description = "Streamlit-based handwritten math grader with OCR and symbolic validation"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "streamlit>=1.28.0",
    "google-generativeai>=0.3.0",
    "sympy>=1.12",
    "opencv-python>=4.8.0",
    "pytesseract>=0.3.10",
    "pillow>=10.0.0",
    "numpy>=1.24.0",
    "python-docx>=0.8.11",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-mock>=3.11.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
]

[tool.black]
line-length = 88
target-version = ['py310']

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
